+++ cookbooks/app-react-context/src/config.ts
@@ -1,13 +1,20 @@
 import type { AppModuleInitiator } from '@equinor/fusion-framework-app';
 import { enableContext } from '@equinor/fusion-framework-react-module-context';
+import buildQuery from 'odata-query';
 
 export const configure: AppModuleInitiator = (configurator) => {
-    enableContext(configurator, {
-        contextType: ['orgchart'], // set contextType to match against
-        /* FIlter result against regex */
-        // contextFilter: (result) => {
-        //     return result.filter((x) => x.title?.match(/ohan/));
-        // },
+    enableContext(configurator, async (builder) => {
+        builder.setContextType(['orgchart']); // set contextType to match against
+        builder.setContextParameterFn(({ search, type }) => {
+            return buildQuery({
+                search,
+                filter: {
+                    type: {
+                        in: type,
+                    },
+                },
+            });
+        });
     });
 };