+++ packages/module-app/src/app/types.ts
@@ -1,0 +1,11 @@
+import type { AppConfig, AppManifest, AppModulesInstance, AppScriptModule } from '../types';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type AppBundleState<TConfig = any, TModules = any> = {
+    appKey: string;
+    status: Set<string>;
+    manifest?: AppManifest;
+    config?: AppConfig<TConfig>;
+    modules?: AppScriptModule;
+    instance?: AppModulesInstance<TModules>;
+};