+++ packages/module-app/src/index.ts
@@ -1,6 +1,14 @@
-export { AppConfigurator, IAppConfigurator, IAppModuleConfig } from './configurator';
-export { AppProvider, IAppProvider } from './provider';
+export {
+    AppModuleConfig,
+    AppConfigurator,
+    IAppConfigurator,
+    AppModuleConfig as IAppModuleConfig,
+} from './AppConfigurator';
+export { AppModuleProvider } from './AppModuleProvider';
 
+export { App } from './app/App';
+
+export * from './events';
 export * from './types';
 
 export { default, AppModule, module, moduleKey } from './module';