+++ packages/module-app/src/module.ts
@@ -1,29 +1,28 @@
-import { Module } from '@equinor/fusion-framework-module';
+import { IModulesConfigurator, Module } from '@equinor/fusion-framework-module';
 import { ModuleDeps } from './types';
 
-import { IAppConfigurator, AppConfigurator } from './configurator';
-import { AppProvider, IAppProvider } from './provider';
+import { IAppConfigurator, AppConfigurator } from './AppConfigurator';
+import { AppModuleProvider } from './AppModuleProvider';
 
 export const moduleKey = 'app';
 
-export type AppModule = Module<
-    typeof moduleKey,
-    IAppProvider,
-    IAppConfigurator<ModuleDeps>,
-    ModuleDeps
->;
+export type AppModule = Module<typeof moduleKey, AppModuleProvider, IAppConfigurator, ModuleDeps>;
 
 export const module: AppModule = {
     name: moduleKey,
     configure: () => new AppConfigurator(),
     initialize: async (args) => {
-        const config = await args.config.createConfig(args);
+        const config = await (args.config as AppConfigurator).createConfig(args);
         const event = await args.requireInstance('event').catch(() => undefined);
-        return new AppProvider({ config, event });
+        return new AppModuleProvider({ config, event });
     },
     dispose: (args) => {
-        (args.instance as unknown as AppProvider).dispose();
+        (args.instance as unknown as AppModuleProvider).dispose();
     },
+};
+
+export const enableAppModule = (configurator: IModulesConfigurator) => {
+    configurator.addConfig({ module });
 };
 
 export default module;