+++ packages/module-context/src/ContextConfigBuilder.ts
@@ -23,6 +23,7 @@ export type ContextConfigBuilderCallback = <TDeps extends Array<AnyModule> = []>
 
 export class ContextConfigBuilder<
     TModules extends Array<AnyModule> = [],
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     TInit extends ModuleInitializerArgs<any, any> = ModuleInitializerArgs<
         ContextModuleConfigurator,
         TModules
@@ -39,6 +40,7 @@ export class ContextConfigBuilder<
 
     requireInstance<T>(module: string): Promise<T>;
 
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     requireInstance(module: string): Promise<any> {
         return this.#init.requireInstance(module);
     }
@@ -49,6 +51,10 @@ export class ContextConfigBuilder<
 
     setContextFilter(filter: ContextModuleConfig['contextFilter']) {
         this.config.contextFilter = filter;
+    }
+
+    setContextParameterFn(fn: ContextModuleConfig['contextParameterFn']) {
+        this.config.contextParameterFn = fn;
     }
 
     setContextClient(