+++ utils/query/src/client/reducer.ts
@@ -1,8 +1,12 @@
+import { original } from 'immer';
+
+import { createReducer as makeReducer } from '@equinor/fusion-observable';
+
 import { QueryClientError } from './QueryClientError';
-import { QueryTaskCompleted, State } from './types';
 
 import { actions } from './actions';
-import { createReducer as makeReducer } from '@equinor/fusion-observable';
+
+import type { QueryTaskCompleted, QueueItem, State } from './types';
 
 export const createReducer = (initial: State = {}) =>
     makeReducer(initial as State, (builder) =>
@@ -45,10 +49,13 @@ export const createReducer = (initial: State = {}) =>
             })
             .addCase(actions.cancel, (state, action) => {
                 const { transaction, reason } = action.payload;
-                const entry = state[transaction];
-                if (entry) {
-                    entry.task.error(
-                        new QueryClientError('abort', 'request was canceled', new Error(reason))
+                if (state[transaction]) {
+                    const request = { ...original(state[transaction]) } as QueueItem;
+                    request.task.error(
+                        new QueryClientError('abort', {
+                            request,
+                            message: reason ?? 'request was canceled',
+                        })
                     );
                     delete state[transaction];
                 }
@@ -60,6 +67,21 @@ export const createReducer = (initial: State = {}) =>
                 if (entry) {
                     entry.errors ??= [];
                     entry.errors.push(payload);
+                }
+            })
+            .addCase(actions.error, (state, action) => {
+                const { payload, meta } = action;
+                const { transaction } = meta.request.meta;
+                if (state[transaction]) {
+                    const request = { ...original(state[transaction]) } as QueueItem;
+                    request.task.error(
+                        new QueryClientError('error', {
+                            request,
+                            message: `failed to process task [${transaction}]`,
+                            cause: payload.cause,
+                        })
+                    );
+                    delete state[transaction];
                 }
             })
     );