+++ cookbooks/app-react-context/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.12](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.11...@equinor/fusion-framework-cookbook-app-react-context@2.1.12) (2022-12-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-cookbook-app-react-context
+
+## [2.1.11](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.10...@equinor/fusion-framework-cookbook-app-react-context@2.1.11) (2022-12-12)
+
+### Bug Fixes
+
+-   **context:** method for contextParameterFn on enableContext ([398658d](https://github.com/equinor/fusion-framework/commit/398658de26355a8ca99aea291963b8c302df3ddc))
+
+## [2.1.10](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.9...@equinor/fusion-framework-cookbook-app-react-context@2.1.10) (2022-12-12)
+
+### Bug Fixes
+
+-   **cookbooks/app-react-context:** update context content ([7120315](https://github.com/equinor/fusion-framework/commit/7120315750543e61e2291e4075b007310b8894fb))
+
+## [2.1.9](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.8...@equinor/fusion-framework-cookbook-app-react-context@2.1.9) (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-cookbook-app-react-context
+
+## [2.1.8](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.7...@equinor/fusion-framework-cookbook-app-react-context@2.1.8) (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-cookbook-app-react-context
+
+## [2.1.7](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.6...@equinor/fusion-framework-cookbook-app-react-context@2.1.7) (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-cookbook-app-react-context
+
+## [2.1.6](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.5...@equinor/fusion-framework-cookbook-app-react-context@2.1.6) (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-cookbook-app-react-context
+
+## [2.1.5](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.4...@equinor/fusion-framework-cookbook-app-react-context@2.1.5) (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-cookbook-app-react-context
+
 ## [2.1.4](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-cookbook-app-react-context@2.1.3...@equinor/fusion-framework-cookbook-app-react-context@2.1.4) (2022-12-06)
 
 **Note:** Version bump only for package @equinor/fusion-framework-cookbook-app-react-context