+++ packages/module-services/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.8](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-services@1.0.7...@equinor/fusion-framework-module-services@1.0.8) (2022-12-12)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-services
+
+## 1.0.7 (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-services
+
+## [1.0.6](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-services@1.0.5...@equinor/fusion-framework-module-services@1.0.6) (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-services
+
+## [1.0.5](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-services@1.0.4...@equinor/fusion-framework-module-services@1.0.5) (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-services
+
+## [1.0.4](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-services@1.0.3...@equinor/fusion-framework-module-services@1.0.4) (2022-12-08)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-services
+
 ## 1.0.3 (2022-12-06)
 
 **Note:** Version bump only for package @equinor/fusion-framework-module-services